modification logo,nom application et profile du admin

+++ routes/web.php
@@ -19,41 +19,54 @@ Route::group(['namespace'=>'Auth'], function () {
 
 
 //les routes pour crud agent
-
-    Route::match(['get', 'head'], 'agents', 'AgentController@index')->name('agents.index')->middleware('admin');
-    Route::post('agents', 'AgentController@store')->name('agents.store')->middleware('admin');
-    Route::match(['get', 'head'], 'agents/create', 'AgentController@create')->name('agents.create');
-    Route::delete('agents/{user}', 'AgentController@destroy')->name('agents.destroy');
-    Route::match(['put', 'path'], 'agents/{user}', 'AgentController@update')->name('agents.update')->middleware('admin');
-    Route::get('agents/{user}/edit', 'AgentController@edit')->name('agents.edit')->middleware('admin');
-    Route::get('agents/{user}/destroy', 'AgentController@destroyForm')->name('agents.delete')->middleware('admin');
-
+Route::group(['namespace'=>'Admin'], function () {
+    Route::match(['get', 'head'], 'admin/agents', 'AgentController@index')->name('agents.index')->middleware('admin');
+    Route::post('admin/agents', 'AgentController@store')->name('agents.store')->middleware('admin');
+    Route::match(['get', 'head'], 'admin/agents/create', 'AgentController@create')->name('agents.create');
+    Route::delete('admin/agents/{user}', 'AgentController@destroy')->name('agents.destroy');
+    Route::match(['put', 'path'], 'admin/agents/{user}', 'AgentController@update')->name('agents.update')->middleware('admin');
+    Route::get('admin/agents/{user}/edit', 'AgentController@edit')->name('agents.edit')->middleware('admin');
+    Route::get('admin/agents/{user}/destroy', 'AgentController@destroyForm')->name('agents.delete')->middleware('admin');
+});
 
 //les routes pour crud bcm
-
-Route::match(['get', 'head'], 'bcm', 'BcmController@index')->name('bcm.index')->middleware('admin');
-Route::post('bcm', 'BcmController@store')->name('bcm.store')->middleware('admin');
-Route::match(['get', 'head'], 'bcm/create', 'BcmController@create')->name('bcm.create');
-Route::delete('bcm/{user}', 'BcmController@destroy')->name('bcm.destroy');
-Route::match(['put', 'path'], 'bcm/{user}', 'BcmController@update')->name('bcm.update')->middleware('admin');
-Route::get('bcm/{user}/edit', 'BcmController@edit')->name('bcm.edit')->middleware('admin');
-Route::get('bcm/{user}/destroy', 'BcmController@destroyForm')->name('bcm.delete')->middleware('admin');
+Route::group(['namespace'=>'Admin'], function () {
+    Route::match(['get', 'head'], 'admin/bcm', 'BcmController@index')->name('bcm.index')->middleware('admin');
+    Route::post('admin/bcm', 'BcmController@store')->name('bcm.store')->middleware('admin');
+    Route::match(['get', 'head'], 'admin/bcm/create', 'BcmController@create')->name('bcm.create');
+    Route::delete('admin/bcm/{user}', 'BcmController@destroy')->name('bcm.destroy');
+    Route::match(['put', 'path'], 'admin/bcm/{user}', 'BcmController@update')->name('bcm.update')->middleware('admin');
+    Route::get('admin/bcm/{user}/edit', 'BcmController@edit')->name('bcm.edit')->middleware('admin');
+    Route::get('admin/bcm/{user}/destroy', 'BcmController@destroyForm')->name('bcm.delete')->middleware('admin');
+});
 
 //les routes pour crud admin
+Route::group(['namespace'=>'Admin'], function () {
+    Route::match(['get', 'head'], 'admin/admin', 'AdminController@index')->name('admin.index')->middleware('admin');
+    Route::post('admin/admin', 'AdminController@store')->name('admin.store')->middleware('admin');
+    Route::match(['get', 'head'], 'admin/admin/create', 'AdminController@create')->name('admin.create');
+    Route::delete('admin/admin/{user}', 'AdminController@destroy')->name('admin.destroy');
+    Route::match(['put', 'path'], 'admin/admin/{user}', 'AdminController@update')->name('admin.update')->middleware('admin');
+    Route::get('admin/admin/{user}/edit', 'AdminController@edit')->name('admin.edit')->middleware('admin');
+    Route::get('admin/admin/{user}/destroy', 'AdminController@destroyForm')->name('admin.delete')->middleware('admin');
+    Route::match(['put', 'path'], 'admin/admi/{user}', 'AdminController@modifProfil')->name('admin.profil')->middleware('admin');
+});
 
-Route::match(['get', 'head'], 'admin', 'AdminController@index')->name('admin.index')->middleware('admin');
-Route::post('admin', 'AdminController@store')->name('admin.store')->middleware('admin');
-Route::match(['get', 'head'], 'admin/create', 'AdminController@create')->name('admin.create');
-Route::delete('admin/{user}', 'AdminController@destroy')->name('admin.destroy');
-Route::match(['put', 'path'], 'admin/{user}', 'AdminController@update')->name('admin.update')->middleware('admin');
-Route::get('admin/{user}/edit', 'AdminController@edit')->name('admin.edit')->middleware('admin');
-Route::get('admin/{user}/destroy', 'AdminController@destroyForm')->name('admin.delete')->middleware('admin');
-
+Route::get('admin', function () {
+    return view('admin');
+});
 
 Route::get('/', function () {
     return redirect('login');
 });
 
+
+//Les routes pour changer le nom et le logo de l'application
+Route::group(['namespace'=>'Admin'], function () {
+    Route::match(['get', 'head'], 'conf', 'ChangerNLController@index')->name('ChangerNL')->middleware('admin');
+    Route::post('conf/changerNom','ChangerNLController@changerNom')->name('changerNom')->middleware('admin');
+    Route::post('conf/changerLogo','ChangerNLController@changerLogo')->name('changerLogo')->middleware('admin');
+});
 
 //Auth::routes();
